requirements.txt
fastapi
uvicorn[standard]
yt-dlp
python-dotenv
google-genai
